+++ src/j1m/relstoragejsonsearch/tests/testupdater.py
@@ -2,11 +2,35 @@ import json
 import mock
 import traceback
 from zope.testing.loggingsupport import InstalledHandler
-from zope.testing.wait import wait
 
 from . import pgbase
 
 class Tests(pgbase.PGTestBase):
+
+    def test_non_empty_generator(self):
+        from ..updater import non_empty_generator
+        self.assertEqual(non_empty_generator(iter(())), None)
+        self.assertEqual(list(non_empty_generator(iter((1, 2, 3)))), [1, 2, 3])
+
+    def test_update_iterator(self):
+        t = 0
+        for i in range(99):
+            if i%7 == 0:
+                t += 1
+            self.store(t, i)
+        from ..updater import Updates
+        import psycopg2
+        conn = psycopg2.connect(self.conn.dsn)
+        self.assertEqual(
+            [[int(r[1]) for r in b]
+             for b in Updates(conn, end_tid=99, limit=20)
+             ],
+            [list(range(0, 21)),
+             list(range(21, 42)),
+             list(range(42, 63)),
+             list(range(63, 84)),
+             list(range(84, 99)),
+             ])
 
     def test_basic(self):
         with mock.patch("j1m.relstoragejsonsearch.updater.logger") as logger:
@@ -17,76 +41,23 @@ class Tests(pgbase.PGTestBase):
             self.store(2, 1, a=1, b=1)
             self.store(2, 2, a=2, b=2)
             self.start_updater()
-            wait((lambda : self.last_tid(2)), 9, message=2)
+            self.wait_tid(2)
             self.store(3, 2, a=3, b=3)
-            wait((lambda : self.last_tid(3)), 9, message=3)
+            self.wait_tid(3)
             self.store(4, 2, a=4, b=4)
-            wait((lambda : self.last_tid(3)), 9, message=4)
+            self.wait_tid(4)
             self.stop_updater()
             self.store(5, 3, n=3)
             self.start_updater()
-            wait((lambda : self.last_tid(5)), 9, 5)
+            self.wait_tid(5)
 
             # Make sure data were stored correctly:
-            self.assertEqual(self.search("""state @> '{"b": 1}'::jsonb"""),
-                             [(1,)])
-            self.assertEqual(self.search("""state @> '{"b": 4}'::jsonb"""),
-                             [(2,)])
-            self.assertEqual(self.search("""state @> '{"n": 3}'::jsonb"""),
-                             [(3,)])
-
-
-    def test_warn_when_no_trigger(self):
-        handler = InstalledHandler(__name__.rsplit('.', 2)[0] + '.updater')
-        self.start_updater()
-        self.store(1, 1, a=1)
-        wait((lambda : self.last_tid(1)), 9, message=1)
-        self.store(2, 1, a=2)
-        wait((lambda : self.last_tid(2)), 9, message=2)
-        self.drop_trigger()
-        self.store(3, 2, a=3)
-        wait((lambda : self.last_tid(3)), 9, message=3)
-        self.assertEqual([(r.msg, r.args) for r in handler.records[2:]],
-                         [('Missed change %s', (3L,))])
-        handler.uninstall()
-
-    def test_catch_up_limit_extra(self):
-        handler = InstalledHandler(__name__.rsplit('.', 2)[0] + '.updater')
-        self.setup_object_json() # so we can drop the trigger :)
-        self.drop_trigger()
-        t=0
-        for o in range(800):
-            if o % 70 == 0:
-                t += 1
-            self.store(t, o, a=o, t=t)
-
-        self.start_updater()
-        wait((lambda : self.last_tid(5)), 9, message=1)
-        self.assertEqual([(r.msg, r.args) for r in handler.records[2:3]],
-                         [('Catch up halting after %s updates, tid %s',
-                           (350, '5'))])
-        wait((lambda : self.last_tid(12)), 9)
-        self.assertEqual([(r.msg, r.args) for r in handler.records[3:4]],
-                         [('Catch up halting after %s updates, tid %s',
-                           (350, '10'))])
-        handler.uninstall()
-
-    def test_catch_up_limit_exact(self):
-        handler = InstalledHandler(__name__.rsplit('.', 2)[0] + '.updater')
-        self.setup_object_json() # so we can drop the trigger :)
-        self.drop_trigger()
-        t=0
-        for o in range(400):
-            if o % 100 == 0:
-                t += 1
-            self.store(t, o, a=o, t=t)
-        self.start_updater()
-        wait((lambda : self.last_tid(4)), 9, message=1)
-        self.assertEqual([(r.msg, r.args) for r in handler.records[2:3]],
-                         [('Catch up halting after %s updates, tid %s',
-                           (300, '3'))])
-        wait((lambda : self.last_tid(4)), 9, message=2)
-        handler.uninstall()
+            self.assertEqual(self.search("""state @> '{"b":
+            1}'::jsonb"""), [(1,)])
+            self.assertEqual(self.search("""state @> '{"b":
+            4}'::jsonb"""), [(2,)])
+            self.assertEqual(self.search("""state @> '{"n":
+            3}'::jsonb"""), [(3,)])
 
     def test_skip_Uninteresting(self):
         import BTrees.OOBTree
@@ -118,6 +89,29 @@ class Tests(pgbase.PGTestBase):
             {u'A': 1,
              u'class_name': u'j1m.relstoragejsonsearch.tests.pgbase.O',
              u'zoid': 1})
+
+    def test_redo(self):
+        # If you change a transformation, you'll want to redo the
+        # updates you made before.  If you use the redo option, then
+        # instead of updating data writen after the tid stored in
+        # object_json_tid, the updater will update the data stored
+        # before, and then stop.
+        self.start_updater()
+        self.store(1, 1, a=1)
+        self.store(2, 2, a=2)
+        self.wait_tid(2)
+        self.stop_updater()
+        self.ex("update object_json_tid set tid=1")
+        from ..updater import main
+        main(['', '-xj1m.relstoragejsonsearch.tests.testupdater:custom',
+              '--redo'])
+        self.ex("select state from object_json order by zoid")
+        self.assertEqual(
+            list(self.cursor),
+            [({u'A': 1,
+               u'class_name': u'j1m.relstoragejsonsearch.tests.pgbase.O',
+               u'zoid': 1},),
+             ({u'a': 2},)])
 
 def custom(zoid, class_name, state):
     state = json.loads(state)